requirements.txt
absl-py==0.7.1
argon2-cffi==20.1.0
astor==0.8.0
async-generator==1.10
attrs==19.3.0
backcall==0.2.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==3.1.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
configparser==4.0.2
contextlib2==0.6.0
cycler==0.10.0
dataclasses==0.7
decorator==4.4.1
defusedxml==0.6.0
easydict==1.9
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
future==0.18.2
gast==0.2.2
GPUtil==1.4.0
grpcio==1.22.0
h5py==2.9.0
idna==2.10
importlib-metadata==0.23
ipaddress==1.0.23
ipydatawidgets==4.0.1
ipykernel==4.10.0
ipyscales==0.3.0
ipython==5.0.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.10.3
joblib==0.14.1
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==5.2.0
jupyter-core==4.6.1
jupyterlab-pygments==0.1.2
K3D==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.3
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==3.0.5
more-itertools==5.0.0
nbclient==0.5.0
nbconvert==5.6.1
nbformat==4.4.0
nest-asyncio==1.4.1
ninja==1.10.0.post2
notebook==5.7.8
numpy==1.19.2
olefile==0.46
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.1
pathlib2==2.3.5
patsy==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.7.1
prompt-toolkit==1.0.15
protobuf==3.9.1
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.4.2
pymesh2 @ https://github.com/PyMesh/PyMesh/releases/download/v0.2.1/pymesh2-0.2.1-cp36-cp36m-linux_x86_64.whl
pyparsing==2.4.0
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.1.0
qtconsole==4.5.5
QtPy==1.9.0
requests==2.24.0
scandir==1.10.0
scikit-learn==0.20.4
scipy==1.2.2
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0
statsmodels==0.10.1
subprocess32==3.5.4
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tflearn==0.3.2
torchfile==0.1.0
tornado==5.1.1
tqdm==4.38.0
traitlets==4.3.3
traittypes==0.2.1
typing-extensions==3.7.4.3
urllib3==1.25.10
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.15.5
widgetsnbextension==3.5.1
zipp==0.6.0
